Update Battlegrid.c
Added in the valid and random ship generation code that was previously in main.c, as well as added code to count the amount of ships in the grid. Comments are added near the end of those pieces of code for further clarification as to the functions of what the code should do.